+++ src/event_loops_bridge.rs
@@ -1,8 +1,13 @@
 use core::cell::RefCell;
 use core::future::Future;
 
+struct WoabRuntime {
+    actix_system_runner: actix::SystemRunner,
+    runtime_cranker_source_id: glib::SourceId,
+}
+
 thread_local! {
-    static ACTIX_SYSTEM_RUNNER: RefCell<actix::SystemRunner> = RefCell::new(actix::System::new());
+    static WOAB_RUNTIME: RefCell<Option<WoabRuntime>> = RefCell::new(None);
 }
 
 /// Run a feature inside the Actix system GTK will be spinning.
@@ -22,9 +27,10 @@ pub fn block_on<F: Future>(fut: F) -> <F as Future>::Output {
 /// using, and instead return the future as the error value so that it could be executed in some
 /// other fashion.
 pub fn try_block_on<F: Future>(fut: F) -> Result<<F as Future>::Output, F> {
-    ACTIX_SYSTEM_RUNNER.with(|system_runner| {
-        if let Ok(system_runner) = system_runner.try_borrow_mut() {
-            let result = system_runner.block_on(fut);
+    WOAB_RUNTIME.with(|woab_runtime| {
+        if let Ok(woab_runtime) = woab_runtime.try_borrow_mut() {
+            let woab_runtime = woab_runtime.as_ref().expect("`try_block_on` called without `run_actix_inside_gtk_event_loop`");
+            let result = woab_runtime.actix_system_runner.block_on(fut);
             Ok(result)
         } else {
             Err(fut)
@@ -33,14 +39,44 @@ pub fn try_block_on<F: Future>(fut: F) -> Result<<F as Future>::Output, F> {
 }
 
 /// Start an Actix `System` that runs inside the GTK thread.
-pub fn run_actix_inside_gtk_event_loop() -> std::io::Result<glib::SourceId> {
-    let source_id = glib::idle_add(|| {
-        try_block_on(async {
-            actix::clock::sleep(core::time::Duration::new(0, 10_000_000)).await;
-        })
-        .map_err(|_| "`idle_add` function called inside Actix context")
-        .unwrap();
-        glib::source::Continue(true)
+pub fn run_actix_inside_gtk_event_loop() -> std::io::Result<()> {
+    WOAB_RUNTIME.with(|woab_runtime| {
+        let mut woab_runtime = woab_runtime.borrow_mut();
+        if woab_runtime.is_some() {
+            panic!("WoAB is already running Actix inside the GTK event loop");
+        }
+        let runtime_cranker_source_id = glib::idle_add(|| {
+            try_block_on(async {
+                actix::clock::sleep(core::time::Duration::new(0, 10_000_000)).await;
+            })
+            .map_err(|_| "`idle_add` function called inside Actix context")
+            .unwrap();
+            glib::source::Continue(true)
+        });
+        *woab_runtime = Some(WoabRuntime {
+            actix_system_runner: actix::System::new(),
+            runtime_cranker_source_id,
+        });
     });
-    Ok(source_id)
+    Ok(())
+}
+
+/// Shut down the Actix `System` that runs inside the GTK thread.
+///
+/// This will close the Actix runtime and stop GTK from idly cranking it to check for new events
+/// from external sources (e.g. network) but will not disconnect the routed GTK signals. If the GTK
+/// loop is still running and these signals are fired, WoAB will panic.
+pub fn close_actix_runtime() -> Result<(), std::io::Error> {
+    WOAB_RUNTIME.with(|woab_runtime| {
+        if let Ok(mut woab_runtime) = woab_runtime.try_borrow_mut() {
+            let woab_runtime = woab_runtime.take().expect("`close_actix_runtime` called before `run_actix_inside_gtk_event_loop`");
+            woab_runtime.actix_system_runner.block_on(async {
+                actix::System::current().stop();
+            });
+            glib::source::source_remove(woab_runtime.runtime_cranker_source_id);
+            woab_runtime.actix_system_runner.run()
+        } else {
+            panic!("`close_actix_runtime` function called inside Actix context");
+        }
+    })
 }